refactor: fix libs dependencies
