+++ libs/wallet-provider/src/index.ts
@@ -1,6 +1,2 @@
 export { useWalletProvider } from './hooks/useWalletProvider'
 export { useWalletChainId } from './hooks/useWalletChainId'
-export { useAccountsLoader } from './hooks/useAccountsLoader'
-export { initWeb3Modal } from './web3ModalConfig'
-export { safeAppsSDK, SAFE_CONNECTOR_UID } from './connectors/safeAppConnector'
-export { hwAccountIndexAtom } from './connectors/trezor/state'
