+++ libs/wallet/src/web3modal/connectors/safeAppConnector.ts
