+++ libs/wallet/src/web3modal/connectors/trezor/TrezorProvider.ts
