+++ libs/wallet/src/web3modal/connectors/trezor/getAccountsList.ts
