+++ libs/wallet/src/web3modal/connectors/trezor/sendTransactionHandler.ts
