+++ libs/wallet/src/web3modal/connectors/trezor/signTypedDataHandler.ts
