+++ libs/wallet/src/web3modal/connectors/trezor/trezorConnector.ts
