+++ libs/wallet/src/web3modal/hooks/useHwAccountIndex.ts
@@ -1,0 +1,6 @@
+import { useAtom } from 'jotai/index'
+import { hwAccountIndexAtom } from '../hwAccountIndexState'
+
+export function useHwAccountIndex() {
+  return useAtom(hwAccountIndexAtom)
+}
